refactor workbench portfolio 360 router
