+++ src/app/routers/workbench.py
@@ -1,9 +1,6 @@
-from fastapi import APIRouter, Path, Query
+from fastapi import APIRouter, Path
 
-from app.contracts.workbench import (
-    WorkbenchOverviewResponse,
-    WorkbenchPortfolio360Response,
-)
+from app.contracts.workbench import WorkbenchOverviewResponse
 from app.middleware.correlation import correlation_id_var
 from app.services.workbench_service_provider import workbench_service
 
@@ -33,35 +30,4 @@ async def get_workbench_overview(
     return await service.get_workbench_overview(
         portfolio_id=portfolio_id,
         correlation_id=correlation_id,
-    )
-
-
-@router.get(
-    "/{portfolio_id}/portfolio-360",
-    response_model=WorkbenchPortfolio360Response,
-    summary="Get Portfolio 360",
-    description=(
-        "Returns the baseline position inventory plus optional projected holdings for an active "
-        "sandbox session. Use this route for live position panels, sandbox comparison, and any "
-        "consumer that needs the same overview context with holdings-level detail attached."
-    ),
-)
-async def get_portfolio_360(
-    portfolio_id: str = Path(
-        ...,
-        description="Canonical portfolio identifier for the portfolio-360 surface.",
-        examples=["PF_1001"],
-    ),
-    session_id: str | None = Query(
-        default=None,
-        description="Optional sandbox session identifier used to overlay projected state.",
-        examples=["sess_1"],
-    ),
-) -> WorkbenchPortfolio360Response:
-    service = workbench_service()
-    correlation_id = correlation_id_var.get()
-    return await service.get_portfolio_360(
-        portfolio_id=portfolio_id,
-        correlation_id=correlation_id,
-        session_id=session_id,
     )
